# ENH: Improve performance of `np.save` for small arrays