Improve performance of `np.save`for small arrays by removing the call to `_filter_header` 
when writing the header, as it is known to be done in Python 3.

---

We use `np.save` to save a lot of small arrays (which are 3D point, so are just 3 ints/floats), 
and we found out that `np.save` is about 20 times slower than `pickle.dumps`.

Profiling shows that most of the time is spent in `_filter_header`, 
which (as I understand it) is not really needed anymore.

The result is about 10x improvement on my machine:
```
import numpy as np
from numpy.lib.format import write_array
import io
array = np.array([2, 3, 4])

%%timeit
buffer = io.BytesIO()
write_array(buffer, array, allow_pickle=False)
```

Before:
`140 µs ± 10.7 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)`

After:
`14.4 µs ± 497 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)`

For reference, using `pickle.dumps` is:
`6.27 µs ± 142 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)`

Edit: This PR also changes reading an array to only call this function for older versions which might have been created by Python 2 before this function was [added in 2014](https://github.com/numpy/numpy/commit/8f068b7866a2959392fae2a15c1a5a19ff79bae9) (which are still used by default),
resulting in a nice improvement as well:

Reading a v1 array (same as today):
`141 µs ± 2.47 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)`

Reading a v3 array now:
~`56 µs ± 13.6 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)`~
`37.3 µs ± 1.11 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)` (after more little improvements)

<!--         ----------------------------------------------------------------
                MAKE SURE YOUR PR GETS THE ATTENTION IT DESERVES!
                ----------------------------------------------------------------

*  FORMAT IT RIGHT:
      http://www.numpy.org/devdocs/dev/development_workflow.html#writing-the-commit-message

*  IF IT'S A NEW FEATURE OR API CHANGE, TEST THE WATERS:
      http://www.numpy.org/devdocs/dev/development_workflow.html#get-the-mailing-list-s-opinion

*  HIT ALL THE GUIDELINES:
      https://numpy.org/devdocs/dev/index.html#guidelines

*  WHAT TO DO IF WE HAVEN'T GOTTEN BACK TO YOU:
      http://www.numpy.org/devdocs/dev/development_workflow.html#getting-your-pr-reviewed
-->