@seberg 2021-03-21
Thanks!  The documentation of `_filter_header` already allures to Python 2 support in reading, so I think that is good as is.

Since this seems worth it... there isn't a version number that we can rely on having stripped headers when reading? I am just curious if we could skip this even on reading most of the time.  IIRC, we do have different version numbers but only use the new one in the rare event that the old version is not able to write the array correctly?  If that is the case, then it seems like its not worth skipping the header cleanup on read.