@ohadravid 2021-03-21
Hi @seberg 👋 
Looking at the release and git history, I think that the header is filtered since 2014, but versions `(1,0)` and `(2,0)` existed before and might have included `L`s.
Versions `(3,0)` comes after 1.16 which was the last release supporting Python 2.7, so we could check `if version >= (3,0)`, but
since `np.save` defaults to saving with the lowest compat this won't help too much I think, but it won't hurt :)