@charris 2021-03-21
This could use a release note. See `doc/release/upcoming_changes/` for examples. If there are potential compatibility problems the release note should probably be in the `compatibility` section, otherwise `improvements`. Maybe two notes if both :)